WIP: AskPath scene
