+++ Assets/Scripts/TES/Components/PathSelectionComponent.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 18f65c96457599848b0d63ff346899ad
-timeCreated: 1475999880
+guid: e9550292f04c2cc4f8e2986e8f5954aa
+timeCreated: 1476386718
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
