--- Assets/Scripts/TES/Components/PathSelectionComponentAlt.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e9550292f04c2cc4f8e2986e8f5954aa
-timeCreated: 1476386718
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
