--- config.ini1
@@ -1,29 +1,0 @@
-# TESUnity Configuration File
-
-[Global]
-PlayMusic = True
-MorrowindPath = D:\Jeux\Oldies\Morrowind
-
-[Rendering]
-RenderPath = 1
-Shader = Bumped
-
-[Lighting]
-AnimateLights = False
-SunShadows = True
-LightShadows = True
-RenderExteriorCellLights = False
-
-[Effects]
-AntiAliasing = False
-AmbientOcclusion = False
-Bloom = True
-WaterBackSideTransparent = False
-
-[VR]
-FollowHeadDirection = True
-DirectModePreview = False
-
-[Debug]
-CreaturesEnabled = True
-
